+++ android/src/main/kotlin/com/sharmadhiraj/installed_apps/InstalledAppsPlugin.kt
@@ -1,30 +1,25 @@
 package com.sharmadhiraj.installed_apps
 
-import android.annotation.SuppressLint
-import android.content.Context
 import android.content.Intent
 import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
-import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
-import android.graphics.Bitmap
-import android.graphics.Bitmap.createBitmap
-import android.graphics.Canvas
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.Drawable
 import android.net.Uri
-import android.os.Build.VERSION.SDK_INT
-import android.os.Build.VERSION_CODES.N_MR1
 import android.provider.Settings.ACTION_APPLICATION_DETAILS_SETTINGS
 import android.widget.Toast
+import android.widget.Toast.LENGTH_LONG
+import android.widget.Toast.LENGTH_SHORT
+import com.sharmadhiraj.installed_apps.Util.Companion.convertAppToMap
+import com.sharmadhiraj.installed_apps.Util.Companion.getContext
+import com.sharmadhiraj.installed_apps.Util.Companion.getPackageManager
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry.Registrar
-import java.io.ByteArrayOutputStream
-
+import java.util.Locale.ENGLISH
 
 class InstalledAppsPlugin(private val registrar: Registrar) : MethodCallHandler {
+
     companion object {
         @JvmStatic
         fun registerWith(registrar: Registrar) {
@@ -34,103 +29,82 @@ class InstalledAppsPlugin(private val registrar: Registrar) : MethodCallHandler 
     }
 
     override fun onMethodCall(call: MethodCall, result: Result) {
-        when {
-            call.method == "getInstalledApps" -> result.success(getInstalledApps())
-            call.method == "startApp" -> {
-                val packageName: String? = call.argument("package_name")
-                startApp(packageName)
+        when (call.method) {
+            "getInstalledApps" -> {
+                val includeSystemApps = call.argument("exclude_system_apps") ?: true
+                val withIcon = call.argument("with_icon") ?: false
+                val packageNamePrefix: String = call.argument("package_name_prefix") ?: ""
+                result.success(getInstalledApps(includeSystemApps, withIcon, packageNamePrefix))
             }
-            call.method == "openSettings" -> {
+            "startApp" -> {
+                val packageName: String? = call.argument("package_name")
+                result.success(startApp(packageName))
+            }
+            "openSettings" -> {
                 val packageName: String? = call.argument("package_name")
                 openSettings(packageName)
+            }
+            "toast" -> {
+                val message = call.argument("message") ?: ""
+                val short = call.argument("short_length") ?: true
+                toast(message, short);
+            }
+            "getAppInfo" -> {
+                val packageName: String = call.argument("package_name") ?: ""
+                result.success(getAppInfo(getPackageManager(registrar), packageName))
+            }
+            "isSystemApp" -> {
+                val packageName: String = call.argument("package_name") ?: ""
+                result.success(isSystemApp(getPackageManager(registrar), packageName))
             }
             else -> result.notImplemented()
         }
     }
 
-    private fun getInstalledApps(): List<Map<String, Any?>> {
-        val packageManager = getPackageManager()
-        val installedApps = packageManager.getInstalledApplications(0)
-        return installedApps
-                .filter { app -> !isSystemApp(app.packageName) }
-                .map { app -> appToMap(packageManager, app) }
+    private fun getInstalledApps(excludeSystemApps: Boolean, withIcon: Boolean, packageNamePrefix: String): List<Map<String, Any?>> {
+        val packageManager = getPackageManager(registrar)
+        var installedApps = packageManager.getInstalledApplications(0)
+        if (excludeSystemApps)
+            installedApps = installedApps.filter { app -> !isSystemApp(packageManager, app.packageName) }
+        if (packageNamePrefix.isNotEmpty())
+            installedApps = installedApps.filter { app -> app.packageName.startsWith(packageNamePrefix.toLowerCase(ENGLISH)) }
+        return installedApps.map { app -> convertAppToMap(packageManager, app, withIcon) }
     }
 
-    private fun appToMap(packageManager: PackageManager, app: ApplicationInfo): HashMap<String, Any?> {
-        val map = HashMap<String, Any?>()
-        map["name"] = packageManager.getApplicationLabel(app)
-        map["package_name"] = app.packageName
-        map["icon"] = drawableToByteArray(app.loadIcon(packageManager))
-        val packageInfo = packageManager.getPackageInfo(app.packageName, 0)
-        map["version_name"] = packageInfo.versionName
-        map["version_code"] = packageInfo.versionCode
-        return map
-    }
-
-    private fun drawableToByteArray(drawable: Drawable): ByteArray {
-        val bitmap = drawableToBitmap(drawable)
-        val stream = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream)
-        return stream.toByteArray()
-    }
-
-    private fun drawableToBitmap(drawable: Drawable): Bitmap {
-        if (SDK_INT <= N_MR1)
-            return (drawable as BitmapDrawable).bitmap
-        val bitmap = createBitmap(drawable.intrinsicWidth, drawable.intrinsicHeight, Bitmap.Config.ARGB_8888)
-        val canvas = Canvas(bitmap)
-        drawable.setBounds(0, 0, canvas.width, canvas.height)
-        drawable.draw(canvas)
-        return bitmap
-    }
-
-    @SuppressLint("NewApi")
-    private fun startApp(packageName: String?) {
-        if (packageName.isNullOrBlank()) {
-            toast("Empty or no package name.")
-            return
-        }
-        try {
-            val appName = getAppName(packageName)
-            toast("Starting app $appName")
-            val launchIntent = getPackageManager().getLaunchIntentForPackage(packageName)
+    private fun startApp(packageName: String?): Boolean {
+        if (packageName.isNullOrBlank()) return false
+        return try {
+            val launchIntent = getPackageManager(registrar).getLaunchIntentForPackage(packageName)
             registrar.context().startActivity(launchIntent)
+            true;
         } catch (e: Exception) {
-            toast("Unable to find app with package name : $packageName")
+            print(e)
+            false
         }
     }
 
-    private fun toast(text: String) {
-        Toast.makeText(getContext(), text, Toast.LENGTH_SHORT).show()
+    private fun toast(text: String, short: Boolean) {
+        Toast.makeText(getContext(registrar), text, if (short) LENGTH_SHORT else LENGTH_LONG).show()
     }
 
-    private fun getAppName(packageName: String?): String {
-        return getPackageManager()
-                .getApplicationLabel(getPackageManager().getApplicationInfo(packageName, PackageManager.GET_META_DATA))
-                .toString()
+    private fun isSystemApp(packageManager: PackageManager, packageName: String): Boolean {
+        return packageManager.getLaunchIntentForPackage(packageName) == null
     }
 
-    private fun getContext(): Context {
-        return registrar.context()
-    }
-
-    private fun getPackageManager(): PackageManager {
-        return getContext().packageManager
-    }
-
-    @SuppressLint("NewApi")
-    private fun isSystemApp(packageName: String): Boolean {
-        return getPackageManager().getLaunchIntentForPackage(packageName) == null
-    }
-
-    @SuppressLint("InlinedApi")
     private fun openSettings(packageName: String?) {
         val intent = Intent()
         intent.flags = FLAG_ACTIVITY_NEW_TASK
         intent.action = ACTION_APPLICATION_DETAILS_SETTINGS
         val uri = Uri.fromParts("package", packageName, null)
         intent.data = uri
-        getContext().startActivity(intent)
+        getContext(registrar).startActivity(intent)
+    }
+
+    private fun getAppInfo(packageManager: PackageManager, packageName: String): Map<String, Any?>? {
+        var installedApps = packageManager.getInstalledApplications(0)
+        installedApps = installedApps.filter { app -> app.packageName == packageName }
+        return if (installedApps.isEmpty()) null
+        else convertAppToMap(packageManager, installedApps[0], true);
     }
 
 }