+++ android/src/main/kotlin/com/sharmadhiraj/installed_apps/Util.kt
@@ -1,0 +1,64 @@
+package com.sharmadhiraj.installed_apps
+
+import android.content.Context
+import android.content.pm.ApplicationInfo
+import android.content.pm.PackageInfo
+import android.content.pm.PackageManager
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import android.graphics.drawable.BitmapDrawable
+import android.graphics.drawable.Drawable
+import android.os.Build.VERSION.SDK_INT
+import android.os.Build.VERSION_CODES.N_MR1
+import android.os.Build.VERSION_CODES.P
+import io.flutter.plugin.common.PluginRegistry
+import java.io.ByteArrayOutputStream
+
+class Util {
+
+    companion object {
+
+        fun convertAppToMap(packageManager: PackageManager, app: ApplicationInfo, withIcon: Boolean): HashMap<String, Any?> {
+            val map = HashMap<String, Any?>()
+            map["name"] = packageManager.getApplicationLabel(app)
+            map["package_name"] = app.packageName
+            map["icon"] = if (withIcon) drawableToByteArray(app.loadIcon(packageManager)) else ByteArray(0)
+            val packageInfo = packageManager.getPackageInfo(app.packageName, 0)
+            map["version_name"] = packageInfo.versionName
+            map["version_code"] = getVersionCode(packageInfo)
+            return map
+        }
+
+        private fun drawableToByteArray(drawable: Drawable): ByteArray {
+            val bitmap = drawableToBitmap(drawable)
+            val stream = ByteArrayOutputStream()
+            bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream)
+            return stream.toByteArray()
+        }
+
+        private fun drawableToBitmap(drawable: Drawable): Bitmap {
+            if (SDK_INT <= N_MR1) return (drawable as BitmapDrawable).bitmap
+            val bitmap = Bitmap.createBitmap(drawable.intrinsicWidth, drawable.intrinsicHeight, Bitmap.Config.ARGB_8888)
+            val canvas = Canvas(bitmap)
+            drawable.setBounds(0, 0, canvas.width, canvas.height)
+            drawable.draw(canvas)
+            return bitmap
+        }
+
+        fun getContext(registrar: PluginRegistry.Registrar): Context {
+            return registrar.context()
+        }
+
+        fun getPackageManager(registrar: PluginRegistry.Registrar): PackageManager {
+            return getContext(registrar).packageManager
+        }
+
+        @Suppress("DEPRECATION")
+        private fun getVersionCode(packageInfo: PackageInfo): Long {
+            return if (SDK_INT < P) packageInfo.versionCode.toLong()
+            else packageInfo.longVersionCode
+        }
+
+    }
+
+}