+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 1.0.0
+## 1.1.0
+*Add getAppInfo & isSystemApp method.
+*Refactor existing methods
 
+## 1.0.0
 *Fixes & Optimization
 
 ## 0.0.1
-
 * Get list of installed app
 * Start app with package name